Cria pastas para arquivos, melhor organização.
