--- GameEngine.py
@@ -1,45 +1,0 @@
-def init_game():
-	"""Iniciliza jogo, recebendo o nome, montando deck e embaralhando cartas
-	"""
-	pass
-	#recebe nome do jogador
-	#monta deck
-	#embaralha deck
-
-def loop_game():
-	"""Loop principal do jogo
-	"""
-	while playing:
-		pass
-		#recebe aposta
-		#distribui cartas
-		#enquanto P1 + 1 carta e P1.pontos < 21
-			#P1 + 1 carta
-		#verifica potuação P1
-			#>21 P1 perde
-		#Dealer verifica pontuação
-			# < 17 : puxa carta
-			#>= 17 : encerra jogada
-
-def take_bet():
-	"""Recebe as apostas
-	"""
-	pass
-
-def hit(deck, hand):
-	"""
-	Retira carta do deck e adciona a mão do jogador atual.
-	Verifica mão < 21
-	
-	Args:
-	    deck (Deck Class): recebe baralho de cartas do jogo atual
-	    hand (Hand Class): recebe mão do jogador
-	"""
-
-def new_game():
-	"""Pergunta se deseja jogar novamente
-	"""
-	pass
-	#pergunta P1 jogar novamente
-		#N : PLAYING = FALSE
-
